state interface for rc refcell
