+++ cw-orch-environment/src/environment/state.rs
@@ -2,7 +2,7 @@
 
 use crate::error::CwEnvError;
 use cosmwasm_std::Addr;
-use std::collections::HashMap;
+use std::{cell::RefCell, collections::HashMap, rc::Rc};
 
 /// State accessor trait.
 /// Indicates that the type has access to an underlying state.
@@ -46,4 +46,34 @@ pub struct DeployDetails {
     pub chain_name: String,
     /// E.g. default
     pub deployment_id: String,
+}
+
+impl<S: StateInterface> StateInterface for Rc<RefCell<S>> {
+    fn get_address(&self, contract_id: &str) -> Result<Addr, CwEnvError> {
+        (**self).borrow().get_address(contract_id)
+    }
+
+    fn set_address(&mut self, contract_id: &str, address: &Addr) {
+        (**self).borrow_mut().set_address(contract_id, address)
+    }
+
+    fn get_code_id(&self, contract_id: &str) -> Result<u64, CwEnvError> {
+        (**self).borrow().get_code_id(contract_id)
+    }
+
+    fn set_code_id(&mut self, contract_id: &str, code_id: u64) {
+        (**self).borrow_mut().set_code_id(contract_id, code_id)
+    }
+
+    fn get_all_addresses(&self) -> Result<HashMap<String, Addr>, CwEnvError> {
+        (**self).borrow().get_all_addresses()
+    }
+
+    fn get_all_code_ids(&self) -> Result<HashMap<String, u64>, CwEnvError> {
+        (**self).borrow().get_all_code_ids()
+    }
+
+    fn deploy_details(&self) -> DeployDetails {
+        (**self).borrow().deploy_details()
+    }
 }
